Ensure log directories exist.
Otherwise it can be very confusing if logs can't be created. It's still
very possible the log dir creation fails due to a lack of permissions,
but at least we try!

If APCu is not available some cycles are wasted on the check, so if
that's a thing you care about, enable APCu in your PHP installation.
Official Docker images enable it.